re-enabled GetObjToNmJson
